pyproject.toml
[project]
name = "article-mind-service"
version = "0.1.0"
description = "Backend service for Article Mind knowledge management system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.2.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    # Content extraction
    "trafilatura>=2.0.0",
    "newspaper4k>=0.9.4",
    "pymupdf>=1.26.0",
    "pymupdf4llm>=0.0.17",
    "playwright>=1.50.0",
    # HTTP and async
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "charset-normalizer>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "bandit>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "trafilatura.*",
    "newspaper.*",
    "fitz.*",
    "pymupdf4llm.*",
    "playwright.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.bandit]
exclude_dirs = ["tests", "alembic/versions"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "alembic/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
